Move RMI helper classes to agent-bootstrap to avoid the need to inject them onto the bootclasspath at runtime
