+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/bootstrap/instrumentation/rmi/ContextPropagator.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.rmi.context;
+package datadog.trace.bootstrap.instrumentation.rmi;
 
 import datadog.trace.bootstrap.ContextStore;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
