+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/bootstrap/instrumentation/rmi/RmiServerDecorator.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.rmi.server;
+package datadog.trace.bootstrap.instrumentation.rmi;
 
 import datadog.trace.bootstrap.instrumentation.api.InternalSpanTypes;
 import datadog.trace.bootstrap.instrumentation.api.UTF8BytesString;
